FFmpegPythonDocker/requirements.txt
azure-storage-blob==1.5.0
numpy==1.14.6